refactor: Refactored code to adhere to MVP format
